third_party/WebKit/LayoutTests/fast/css/deprecated-flex-box-zero-width-intrinsic-max-width-expected.txt

CONSOLE WARNING: The 'intrinsic' CSS keyword is deprecated and will be removed in M47, around November 2015. Use the standard 'max-content' keyword instead.
Tests if the Deprecated Flex boxes with max-width: intrinsic and width:0 has non-zero calculated width

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS flexBoxDiv.width is non-zero.
PASS successfullyParsed is true

TEST COMPLETE
